+++ prisma/migrations/20240606183122_simplify_the_schema/migration.sql
@@ -1,0 +1,42 @@
+-- CreateTable
+CREATE TABLE "Payment" (
+    "id" TEXT NOT NULL,
+    "rut" TEXT,
+    "price" DOUBLE PRECISION NOT NULL,
+    "ticketAmount" INTEGER NOT NULL,
+    "date" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "signature_token" TEXT,
+    "payment_status" TEXT,
+    "payment_token" TEXT,
+    "customer_email" TEXT NOT NULL,
+    "customer_name" TEXT NOT NULL,
+    "customer_phone" TEXT,
+    "productId" TEXT NOT NULL,
+    "buys" JSONB NOT NULL,
+    "payment_id_service" TEXT,
+    "ticketsType" TEXT NOT NULL,
+
+    CONSTRAINT "Payment_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateTable
+CREATE TABLE "Product" (
+    "id" TEXT NOT NULL,
+    "name" TEXT NOT NULL,
+    "stock" INTEGER,
+    "date" TIMESTAMPTZ(6),
+
+    CONSTRAINT "Product_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Payment_id_key" ON "Payment"("id");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Payment_payment_id_service_key" ON "Payment"("payment_id_service");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Product_id_key" ON "Product"("id");
+
+-- AddForeignKey
+ALTER TABLE "Payment" ADD CONSTRAINT "Payment_productId_fkey" FOREIGN KEY ("productId") REFERENCES "Product"("id") ON DELETE RESTRICT ON UPDATE CASCADE;