+++ src/routes/(app)/+page.server.ts
@@ -10,13 +10,7 @@ import {
 } from "$lib/config/sanity/queries";
 import { error, json, redirect } from "@sveltejs/kit";
 import type { Actions, PageServerLoad } from "./$types";
-import { PRIVATE_TOKEN_PAYKU } from "$env/static/private";
-import {
-  PUBLIC_TOKEN_PAYKU,
-  PUBLIC_PAYKU_API_URL,
-  PUBLIC_PAYMENT_COMPLETED_URL,
-  PUBLIC_PAYMENT_WEBHOOK_URL_PAYKU,
-} from "$env/static/public";
+import { PRIVATE_TOKEN_PAYKU, VITE_PAYKU_API_URL, VITE_PAYMENT_COMPLETED_URL, VITE_PAYMENT_WEBHOOK_URL_PAYKU, VITE_PUBLIC_TOKEN_PAYKU } from "$env/static/private";
 import { client } from "$lib/server/prisma";
 import { calculatePrice } from "$lib/utils/eventUtils";
 
@@ -173,8 +167,8 @@ export const actions: Actions = {
       subject: `${tickets} entradas para ${nextEvent.title}`,
       name: name,
       country: "Chile",
-      urlreturn: PUBLIC_PAYMENT_COMPLETED_URL,
-      urlnotify: PUBLIC_PAYMENT_WEBHOOK_URL_PAYKU,
+      urlreturn: VITE_PAYMENT_COMPLETED_URL,
+      urlnotify: VITE_PAYMENT_WEBHOOK_URL_PAYKU,
       payment: 1,
       additional_parameters: {
         event_id: nextEvent._id,
@@ -184,11 +178,11 @@ export const actions: Actions = {
     let dataUrlRedirect = "";
 
     try {
-      const response = await fetch(`${PUBLIC_PAYKU_API_URL}/transaction`, {
+      const response = await fetch(`${VITE_PAYKU_API_URL}/transaction`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
-          Authorization: `Bearer ${PUBLIC_TOKEN_PAYKU}`,
+          Authorization: `Bearer ${VITE_PUBLIC_TOKEN_PAYKU}`,
         },
         body: JSON.stringify(payload),
       });
@@ -283,8 +277,8 @@ export const actions: Actions = {
       subject: `${tickets} entradas para ${nextEvent.title}`,
       name: name,
       country: "Chile",
-      urlreturn: PAYMENT_COMPLETED_URL,
-      urlnotify: PAYMENT_WEBHOOK_URL_PAYKU,
+      urlreturn: VITE_PAYMENT_COMPLETED_URL,
+      urlnotify: VITE_PAYMENT_WEBHOOK_URL_PAYKU,
       additional_parameters: {
         event_id: nextEvent._id,
       },
@@ -297,7 +291,7 @@ export const actions: Actions = {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
-          Authorization: `Bearer ${PUBLIC_TOKEN_PAYKU}`,
+          Authorization: `Bearer ${VITE_PUBLIC_TOKEN_PAYKU}`,
         },
         body: JSON.stringify(payload),
       });
@@ -322,7 +316,6 @@ export const actions: Actions = {
         },
       });
 
-      console.log(payment);
       if (payment.payment_id_service) {
         event.cookies.set("payment_id_service", payment?.payment_id_service, {
           // send cookie for every page